# config/openapi.yaml
...
openapi: 3.1.0
info:
  title: SASO REST API
  version: 1.0.0-alpha
  summary: Inventory and warehouse operations over HTTP.
  description: |
    REST surface for the SASO inventory and warehouse management system.
    Errors follow [RFC 7807 Problem Details](https://datatracker.ietf.org/doc/html/rfc7807)
    with the SASO `code` and `traceId` vendor extensions. See the
    [error codes catalogue](https://docs.willen-federation.org/error-codes)
    for the full identifier list.
  license:
    name: GPL-3.0-or-later
    url: https://www.gnu.org/licenses/gpl-3.0.html
  contact:
    name: Willen Federation contributors
    url: https://github.com/Willen-Federation/SASO-Willen-Edition
servers:
  - url: /api/v1
    description: This SASO instance.
...
paths:
...
  /mobile/connect:
    post:
      tags: [mobile]
      summary: Exchange a pairing code for a JWT token pair.
      description: |
        The Flutter app calls this automatically after scanning the QR and
        detecting the `SASO1:` prefix. On success the pairing code is consumed
        (single use) and an OAuth2-style token pair is returned:

        - `access_token` — HS256 JWT, valid 1 hour, use as `Authorization: Bearer`
        - `refresh_token` — opaque long-lived token (1 year), use to renew the JWT
          via `POST /api/v1/mobile/token/refresh`

        Store both tokens securely (Flutter secure_storage or equivalent). The
        refresh token is rotated on every refresh call.
      operationId: mobileConnect
      x-handler: 'Saso\Presentation\Api\V1\Controller\Mobile\ConnectController::handle'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRequest'
      responses:
        '201':
          description: JWT token pair issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '400':
          $ref: '#/components/responses/ProblemBadRequest'
        '404':
          $ref: '#/components/responses/ProblemNotFound'
        '500':
          $ref: '#/components/responses/ProblemInternalError'
...
components:
  schemas:
...
    ConnectRequest:
      type: object
      required: [token, deviceName]
      properties:
        token:
          type: string
          description: |
            Raw token extracted from the QR payload (the segment between
            `SASO1:` and `|`).
          example: dGhpcyBpcyBhIHRlc3Q
        deviceName:
          type: string
          maxLength: 200
          description: Human-readable name the Flutter app supplies for this device.
          example: iPad mini
...
    TokenPairResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token, device_id, expires_at]
      description: OAuth2-style token response (RFC 6749 §5.1).
      properties:
        access_token:
          type: string
          description: |
            HS256 JWT access token (valid for `expires_in` seconds).
            Include as `Authorization: Bearer <access_token>` on API calls.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds (3600 = 1 hour).
          example: 3600
        refresh_token:
          type: string
          description: |
            Long-lived opaque refresh token (valid ~1 year). Store securely
            and use to obtain a new access token via POST /mobile/token/refresh.
            **Rotated on every refresh** — discard the old token after each call.
          example: Y0zM3nOpQ4rT5uV6wX7y
        device_id:
          type: integer
          description: Numeric device record ID (use for admin revocation).
          example: 1
        device_name:
          type: string
          description: Device name as supplied during pairing.
          example: iPad mini
        expires_at:
          type: string
          format: date-time
          description: Access token expiry (RFC 3339).
          example: '2026-04-26T13:00:00Z'
...
    ProblemDetails:
      type: object
      description: |
        RFC 7807 Problem Details with the SASO vendor extensions `code` and
        `traceId`. Clients should branch on `code`; `title` and `detail` are
        localised strings whose wording may change between releases.
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
          description: Canonical URL for the error code.
          example: https://docs.willen-federation.org/error-codes#SASO-AUTH-1001
        title:
          type: string
          description: Short, human-readable summary.
          example: Invalid credentials
        status:
          type: integer
          description: Mirrors the HTTP status of the response.
          example: 401
        detail:
          type: string
          description: Request-specific message, localised.
          example: The submitted credentials did not match an active account.
        instance:
          type: string
          description: Request URI (no query string).
          example: /api/v1/auth/login
        code:
          type: string
          pattern: '^SASO-[A-Z]+-\d{4}$'
          description: Stable SASO error identifier.
          example: SASO-AUTH-1001
        traceId:
          type: string
          format: uuid
          description: UUIDv4 unique per request, present in the server log.
          example: 1f9b3c8a-9e15-4d6a-8c2c-3d8f4f1f7a12
  responses:
...
    ProblemBadRequest:
      description: The request body is missing or invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ProblemNotFound:
      description: No matching endpoint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ProblemInternalError:
      description: Unexpected server-side failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'